Inset types moved to an external file
